+++ packages/article/showcase-helper.js
@@ -12,6 +12,7 @@ import {
   MockedProvider,
   schemaToMocks
 } from "@times-components/provider-test-tools";
+import saveArticleApi from "@times-components/save-star-web/mock-save-api-showcase";
 import { sections } from "@times-components/storybook";
 import { scales, themeFactory } from "@times-components/styleguide";
 import storybookReporter from "@times-components/tealium-utils";
@@ -211,6 +212,7 @@ const renderArticle = ({
   id,
   inDepthBackgroundColour,
   inDepthTextColour,
+  saveApi = saveArticleApi,
   scale,
   section,
   template,
@@ -275,6 +277,7 @@ const renderArticle = ({
               onVideoPress={preventDefaultedAction(decorateAction)(
                 "onVideoPress"
               )}
+              saveApi={saveApi}
               refetch={refetch}
             />
           </MessageManager>
@@ -370,6 +373,7 @@ const renderArticleConfig = ({
             inDepthTextColour,
             isTeaser: isTeaser || withTeaser,
             isMeteredExpired: isMeteredExpired || isMeteredExpiredPage,
+            saveApi: saveArticleApi,
             scale,
             section,
             template