+++ packages/edition-slices/src/tiles/tile-v/styles/index.js
@@ -3,7 +3,8 @@ import { fonts, spacing } from "@times-components/styleguide";
 const styles = {
   container: {
     paddingHorizontal: spacing(2),
-    paddingTop: spacing(2)
+    paddingTop: spacing(2),
+    flex: 1
   },
   headline: {
     fontFamily: fonts.headline,