+++ ios-app/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.164](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.7.163...@times-components/ios-app@0.7.164) (2019-05-24)
+
+**Note:** Version bump only for package @times-components/ios-app
+
+
+
+
+
+## [0.7.163](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.7.162...@times-components/ios-app@0.7.163) (2019-05-24)
+
+**Note:** Version bump only for package @times-components/ios-app
+
+
+
+
+
+## [0.7.162](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.7.161...@times-components/ios-app@0.7.162) (2019-05-24)
+
+**Note:** Version bump only for package @times-components/ios-app
+
+
+
+
+
+## [0.7.161](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.7.160...@times-components/ios-app@0.7.161) (2019-05-23)
+
+**Note:** Version bump only for package @times-components/ios-app
+
+
+
+
+
+## [0.7.160](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.7.159...@times-components/ios-app@0.7.160) (2019-05-23)
+
+**Note:** Version bump only for package @times-components/ios-app
+
+
+
+
+
+## [0.7.159](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.7.158...@times-components/ios-app@0.7.159) (2019-05-23)
+
+**Note:** Version bump only for package @times-components/ios-app
+
+
+
+
+
 ## [0.7.158](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.7.157...@times-components/ios-app@0.7.158) (2019-05-22)
 
 **Note:** Version bump only for package @times-components/ios-app