+++ packages/article-byline/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.1.73](https://github.com/newsuk/times-components/compare/@times-components/article-byline@3.1.72...@times-components/article-byline@3.1.73) (2019-05-23)
+
+**Note:** Version bump only for package @times-components/article-byline
+
+
+
+
+
 ## [3.1.72](https://github.com/newsuk/times-components/compare/@times-components/article-byline@3.1.71...@times-components/article-byline@3.1.72) (2019-05-22)
 
 **Note:** Version bump only for package @times-components/article-byline