+++ packages/pages/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.49](https://github.com/newsuk/times-components/compare/@times-components/pages@2.0.48...@times-components/pages@2.0.49) (2019-05-24)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
+## [2.0.48](https://github.com/newsuk/times-components/compare/@times-components/pages@2.0.47...@times-components/pages@2.0.48) (2019-05-24)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
+## [2.0.47](https://github.com/newsuk/times-components/compare/@times-components/pages@2.0.46...@times-components/pages@2.0.47) (2019-05-24)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
+## [2.0.46](https://github.com/newsuk/times-components/compare/@times-components/pages@2.0.45...@times-components/pages@2.0.46) (2019-05-23)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
+## [2.0.45](https://github.com/newsuk/times-components/compare/@times-components/pages@2.0.44...@times-components/pages@2.0.45) (2019-05-23)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
+## [2.0.44](https://github.com/newsuk/times-components/compare/@times-components/pages@2.0.43...@times-components/pages@2.0.44) (2019-05-23)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
 ## [2.0.43](https://github.com/newsuk/times-components/compare/@times-components/pages@2.0.42...@times-components/pages@2.0.43) (2019-05-22)
 
 **Note:** Version bump only for package @times-components/pages