+++ packages/slice-layout/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.29.5](https://github.com/newsuk/times-components/compare/@times-components/slice-layout@0.29.4...@times-components/slice-layout@0.29.5) (2019-05-23)
+
+**Note:** Version bump only for package @times-components/slice-layout
+
+
+
+
+
 ## [0.29.4](https://github.com/newsuk/times-components/compare/@times-components/slice-layout@0.29.3...@times-components/slice-layout@0.29.4) (2019-05-22)
 
 **Note:** Version bump only for package @times-components/slice-layout